threshold: Fix cases when landing on thresholds
Background:
For non-range thresholds, the thresholds are exclusive. For range
thresholds, the thresholds are inclusive. That is:

When upper is set, the state is true for (upper, ∞].
When lower is set, the state is true for [-∞, lower).
When upper and lower are both set, the state is true for [lower, upper].

The logic wasn't handling all cases properly from all directions.
Specifically, for upper sensors, transitioning from an above value to
the threshold was incorrect. For lower sensors, transitioning from a
below value to the threshold was incorrect.

Practically speaking, this means that a sensor measuring power
with an upper threshold of 0 would never become false when the
power was turned off.

The fixes also include the corresponding behavior changes when using
hysteresis.